requirements.txt
beautifulsoup4==4.13.3
clip==1.0
ImageHash==4.3.2
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.29.0
torch==2.6.0
